+++ include/matador/db/mysql/mysql_prepared_result.hpp
@@ -68,9 +68,9 @@ protected:
   void read_value(const char *id, size_type index, double &value) override;
   void read_value(const char *id, size_type index, matador::time &value) override;
   void read_value(const char *id, size_type index, matador::date &value) override;
-  void read_value(const char *id, size_type index, char *value, long size) override;
+  void read_value(const char *id, size_type index, char *value, size_t size) override;
   void read_value(const char *id, size_type index, std::string &value) override;
-  void read_value(const char *id, size_type index, std::string &value, long size) override;
+  void read_value(const char *id, size_type index, std::string &value, size_t size) override;
 
 protected:
   bool needs_bind() override;
